+++ cmd/cli/command/deployment/deployment.go
@@ -11,5 +11,7 @@ var Command = cli.Command{
 	Usage:   "Manage your Common Fate Deployment",
 	Subcommands: []*cli.Command{
 		&logs.Command,
+		&diagnosticsCommand,
+		&backgroundJobsCommand,
 	},
 }